Add BaseFlyingEnemy file
